move up public SearchScreenTestTag
