+++ app/src/androidTest/java/ly/david/mbjc/ui/search/SearchEachEntityTest.kt
@@ -22,7 +22,7 @@ import ly.david.data.test.toFakeMusicBrainzModel
 import ly.david.mbjc.MainActivityTest
 import ly.david.mbjc.StringReferences
 import ly.david.mbjc.ui.TopLevelScaffold
-import ly.david.musicsearch.feature.search.internal.SearchScreenTestTag
+import ly.david.musicsearch.feature.search.SearchScreenTestTag
 import ly.david.ui.common.getDisplayTextRes
 import ly.david.ui.core.theme.PreviewTheme
 import org.junit.Before
