+++ feature/search/src/main/java/ly/david/musicsearch/feature/search/SearchScreenTestTag.kt
@@ -1,0 +1,5 @@
+package ly.david.musicsearch.feature.search
+
+enum class SearchScreenTestTag {
+    TEXT_FIELD,
+}
